chore: split events exception utility functions
