+++ src/events/exceptions/index.ts
@@ -1,3 +1,4 @@
 export * from './exceptions';
 export * from './exceptions.interfaces';
 export * from './exceptions.examples';
+export * from './exceptions.utils';
